Parser: fix for test case n_array_a_invalid_utf8.
The following happens in awk:

```sh
# printf "a\xe5" | gawk --posix '{print length($0); gsub(/./,"(&)"); print}'
gawk: cmd. line:1: (FILENAME=- FNR=1) warning: Invalid multibyte data detected.
There may be a mismatch between your data and your locale.
2
(a)
```

So, even though awk reports string length 2, the /./ pattern in gsub
can't match \xe5.  Now, consider how this impacts the tokenizer, whose
pattern ends with `|.`, which should lead to replacing each
yet-unmatched input character:

```
Input JSON(*)      Tokens
[a]                [, a, ]
[ab]               [, a, b, ]
[a`<e5>`]          [, a`<e5>`, ]
```
The last line is a problem, because, before this commit, `parse_value()`
assumed that any token corresponding to yet-unmatched input has length 1.
Consequently, the parser recognized a`<e5>` as a <value> token.
To fix this case, we add `is_value()` to explicitly test that the
unmatched input token is indeed a well-formed <value> token.(**)

(*) The last line is test case `n_array_a_invalid_utf8.json` of the
[JSONTestSuite](https://github.com/nst/JSONTestSuite.git)
In an ideal language, the parser shouldn't need to test that tokens are
well-formed -- that's the tokenizer's concern.

(**) Ideally, the parser shouldn't be concerned with a token's form,
which should be validated by the tokenizer exclusively. However, the
tokenizer is constrained by awk's regex engine implementation.